requirements.txt
cssselect==0.9.2
Jinja2==2.8
lxml==4.3.0
Markdown==2.6.6
MarkupSafe==0.23
olefile==0.43
Pillow==5.4.1
pip-autoremove==0.9.0
pyquery==1.2.13
python-dateutil==2.6.0
python-frontmatter==0.3.1
PyYAML==4.2b4
requests>=2.20.0
six==1.10.0
